Migrated out pydantic models to keep api.py clean and focused primarily on API logic. Pydantic models will now be located under /pydantic_models nad imported as from pydantic_models import *
